🚚 将接口移动到 `fetcher` 下
