+++ src/router/front_end/config/datasource.rs
@@ -1,7 +1,4 @@
-use axum::{
-    routing::{get, post},
-    Router,
-};
+use axum::{routing::get, Router};
 
 use crate::router::ServerRoute;
 
@@ -10,8 +7,4 @@ pub struct ConfigDatasourceFrontend;
 pub(super) fn datasource_router() -> ServerRoute {
     Router::new()
         .route("/list", get(ConfigDatasourceFrontend::datasource_list))
-        .route(
-            "/standaloneFetcherConfig/:combine_id",
-            post(ConfigDatasourceFrontend::standalone_fetcher_config),
-        )
 }
