+++ src/router/front_end/config/fetcher.rs
@@ -1,0 +1,12 @@
+use axum::{routing::get, Router};
+
+pub struct ConfigFetcherFrontend;
+
+use crate::router::ServerRoute;
+
+pub fn fetcher_router() -> ServerRoute {
+    Router::new().route(
+        "/standaloneConfig/:combine_id",
+        get(ConfigFetcherFrontend::standalone_fetcher_config),
+    )
+}
