+++ src/router/front_end/config/mod.rs
@@ -1,10 +1,14 @@
 use axum::Router;
 
-use self::datasource::datasource_router;
+use self::{datasource::datasource_router, fetcher::fetcher_router};
 use crate::router::ServerRoute;
 pub mod datasource;
-pub use datasource::ConfigDatasourceFrontend;
+pub mod fetcher;
+pub use self::datasource::ConfigDatasourceFrontend;
+pub use self::fetcher::ConfigFetcherFrontend;
 
 pub(super) fn config_router() -> ServerRoute {
-    Router::new().nest("/datasource", datasource_router())
+    Router::new()
+        .nest("/datasource", datasource_router())
+        .nest("/fetcher", fetcher_router())
 }
