+++ src/serves/frontend/config/fetcher/mod.rs
@@ -1,0 +1,2 @@
+pub mod controllers;
+mod error;
