+++ src/serves/frontend/config/mod.rs
@@ -1,1 +1,2 @@
 pub mod datasource;
+pub mod fetcher;
